+++ core/src/main/java/com/asml/apa/wta/core/config/RuntimeConfig.java
@@ -29,5 +29,8 @@ public class RuntimeConfig {
   @Builder.Default
   private String logLevel = "ERROR";
 
+  @Builder.Default
+  private boolean isStageLevel = false;
+
   private String outputPath;
 }